# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Currency Exchange API
  version: 1.0.0
  description: |
    OpenAPI contract for the currency exchange use cases:
    - convert one amount to one target currency,
    - convert one amount to multiple target currencies,
    - fetch historical rates for a given date.
servers:
  - url: http://localhost:8080
    description: Local environment

tags:
  - name: Exchange
    description: Currency conversion and historical rates

paths:
  /api/v1/exchange/conversion:
    get:
      tags:
        - Exchange
      summary: Get all conversions consulted
      operationId: getAllConversions
      responses:
        '200':
          description: List of all consulted conversions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversion'
              examples:
                success:
                  value:
                    - date: '2026-05-01'
                      from:
                        currency: EUR
                        amount: 100.00
                      to:
                        currency: USD
                        amount: 102.00
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Exchange
      summary: Convert amount to one target currency
      operationId: convertToSingleCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleConversionRequest'
            examples:
              success:
                value:
                  from: EUR
                  to: USD
                  amount: 100
      responses:
        '200':
          description: Conversion completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertedAmount'
              examples:
                success:
                  value:
                    currency: USD
                    amount: 102.00
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/exchange/convert/bulk:
    get:
      tags:
        - Exchange
      summary: Convert amount to multiple target currencies
      operationId: convertToMultipleCurrencies
      parameters:
        - $ref: '#/components/parameters/fromCurrency'
        - $ref: '#/components/parameters/amount'
        - $ref: '#/components/parameters/targetCurrencies'
      responses:
        '200':
          description: Conversion completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkConversionResponse'
              examples:
                success:
                  value:
                    results:
                      - currency: USD
                        amount: 102.00
                      - currency: GBP
                        amount: 78.40
                      - currency: JPY
                        amount: 15300.00
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/exchange/historical:
    get:
      tags:
        - Exchange
      summary: Get historical rates for a specific date
      operationId: getHistoricalRates
      parameters:
        - $ref: '#/components/parameters/fromCurrency'
        - $ref: '#/components/parameters/amount'
        - $ref: '#/components/parameters/targetCurrencies'
        - $ref: '#/components/parameters/date'
      responses:
        '200':
          description: Historical rates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalRateResponse'
              examples:
                success:
                  value:
                    date: '2026-05-01'
                    exchangeRateResponse:
                      data:
                        USD: 1.02
                        GBP: 0.78
                        JPY: 153.00
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    fromCurrency:
      name: from
      in: query
      required: true
      description: Base currency of the provided amount
      schema:
        $ref: '#/components/schemas/Currency'

    toCurrency:
      name: to
      in: query
      required: true
      description: Target currency for single conversion
      schema:
        $ref: '#/components/schemas/Currency'

    targetCurrencies:
      name: targets
      in: query
      required: true
      description: Comma-separated target currencies for bulk conversion and historical rates
      schema:
        type: array
        minItems: 1
        items:
          $ref: '#/components/schemas/Currency'
      style: form
      explode: false
      example: EUR,USD,JPY

    amount:
      name: amount
      in: query
      required: true
      description: Monetary amount to convert
      schema:
        type: number
        format: double
        exclusiveMinimum: 0
      example: 100

    date:
      name: date
      in: query
      required: true
      description: Historical date in ISO-8601 format (yyyy-MM-dd)
      schema:
        type: string
        format: date
      example: '2026-05-01'

  responses:
    BadRequest:
      description: Invalid input provided by the client
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCurrency:
              value:
                code: BAD_REQUEST
                message: Invalid currency code
            missingParameter:
              value:
                code: BAD_REQUEST
                message: Required request parameter is missing

    InternalServerError:
      description: Unexpected error while processing the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unexpectedError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: Unexpected internal error

  schemas:
    SingleConversionRequest:
      type: object
      required:
        - from
        - to
        - amount
      properties:
        from:
          $ref: '#/components/schemas/Currency'
        to:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          format: double
          exclusiveMinimum: 0
          example: 100

    Currency:
      type: string
      enum:
        - EUR
        - USD
        - JPY
        - BGN
        - CZK
        - DKK
        - GBP
        - HUF
        - PLN
        - RON
        - SEK
        - CHF
        - ISK
        - NOK
        - HRK
        - RUB
        - TRY
        - AUD
        - BRL
        - CAD
        - CNY
        - HKD
        - IDR
        - ILS
        - INR
        - KRW
        - MXN
        - MYR
        - NZD
        - PHP
        - SGD
        - THB
        - ZAR

    ConvertedAmount:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          format: double
          description: Converted amount rounded to 2 decimals
          example: 102.0

    MoneyAmount:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          type: number
          format: double
          description: Monetary amount rounded to 2 decimals
          example: 100.00

    Conversion:
      type: object
      required:
        - date
        - from
        - to
      properties:
        date:
          type: string
          format: date
          example: '2026-05-01'
        from:
          $ref: '#/components/schemas/MoneyAmount'
        to:
          $ref: '#/components/schemas/MoneyAmount'

    BulkConversionResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ConvertedAmount'

    ExchangeRateResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Exchange rate values indexed by target currency code
          example:
            USD: 1.02
            GBP: 0.78
            JPY: 153.0

    HistoricalRateResponse:
      type: object
      required:
        - date
        - exchangeRateResponse
      properties:
        date:
          type: string
          format: date
          example: '2026-05-01'
        exchangeRateResponse:
          $ref: '#/components/schemas/ExchangeRateResponse'

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Invalid request parameters