# target/generated-sources/openapi/src/main/java/edu/itba/class10/ExchangeApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package edu.itba.class10;

import edu.itba.class10.converter.api.domain.BulkConversionResponse;
import edu.itba.class10.converter.api.domain.Conversion;
import edu.itba.class10.converter.api.domain.ConvertedAmount;
import edu.itba.class10.converter.api.domain.Currency;
import org.springframework.format.annotation.DateTimeFormat;
import edu.itba.class10.converter.api.domain.ErrorResponse;
import edu.itba.class10.converter.api.domain.HistoricalRateResponse;
import java.time.LocalDate;
import edu.itba.class10.converter.api.domain.SingleConversionRequest;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-06-11T14:38:46.369826200-03:00[America/Buenos_Aires]", comments = "Generator version: 7.13.0")
@Validated
public interface ExchangeApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /api/v1/exchange/convert/bulk : Convert amount to multiple target currencies
     *
     * @param from Base currency of the provided amount (required)
     * @param amount Monetary amount to convert (required)
     * @param targets Comma-separated target currencies for bulk conversion and historical rates (required)
     * @return Conversion completed successfully (status code 200)
     *         or Invalid input provided by the client (status code 400)
     *         or Unexpected error while processing the request (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/exchange/convert/bulk",
        produces = { "application/json" }
    )
    
    default ResponseEntity<BulkConversionResponse> convertToMultipleCurrencies(
        @NotNull  @Valid @RequestParam(value = "from", required = true) Currency from,
        @NotNull  @Valid @RequestParam(value = "amount", required = true) Double amount,
        @NotNull @Size(min = 1)  @Valid @RequestParam(value = "targets", required = true) List<Currency> targets
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"results\" : [ { \"amount\" : 102.0, \"currency\" : \"EUR\" }, { \"amount\" : 102.0, \"currency\" : \"EUR\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/v1/exchange/conversion : Convert amount to one target currency
     *
     * @param singleConversionRequest  (required)
     * @return Conversion completed successfully (status code 200)
     *         or Invalid input provided by the client (status code 400)
     *         or Unexpected error while processing the request (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/exchange/conversion",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ConvertedAmount> convertToSingleCurrency(
         @Valid @RequestBody SingleConversionRequest singleConversionRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"amount\" : 102.0, \"currency\" : \"EUR\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/exchange/conversion : Get all conversions consulted
     *
     * @return List of all consulted conversions (status code 200)
     *         or Unexpected error while processing the request (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/exchange/conversion",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Conversion>> getAllConversions(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"date\" : \"2026-05-01\", \"from\" : { \"amount\" : 100.0, \"currency\" : \"EUR\" }, \"to\" : { \"amount\" : 100.0, \"currency\" : \"EUR\" } }, { \"date\" : \"2026-05-01\", \"from\" : { \"amount\" : 100.0, \"currency\" : \"EUR\" }, \"to\" : { \"amount\" : 100.0, \"currency\" : \"EUR\" } } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/v1/exchange/historical : Get historical rates for a specific date
     *
     * @param from Base currency of the provided amount (required)
     * @param amount Monetary amount to convert (required)
     * @param targets Comma-separated target currencies for bulk conversion and historical rates (required)
     * @param date Historical date in ISO-8601 format (yyyy-MM-dd) (required)
     * @return Historical rates retrieved successfully (status code 200)
     *         or Invalid input provided by the client (status code 400)
     *         or Unexpected error while processing the request (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/exchange/historical",
        produces = { "application/json" }
    )
    
    default ResponseEntity<HistoricalRateResponse> getHistoricalRates(
        @NotNull  @Valid @RequestParam(value = "from", required = true) Currency from,
        @NotNull  @Valid @RequestParam(value = "amount", required = true) Double amount,
        @NotNull @Size(min = 1)  @Valid @RequestParam(value = "targets", required = true) List<Currency> targets,
        @NotNull  @Valid @RequestParam(value = "date", required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"date\" : \"2026-05-01\", \"exchangeRateResponse\" : { \"data\" : { \"USD\" : 1.02, \"GBP\" : 0.78, \"JPY\" : 153.0 } } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"BAD_REQUEST\", \"message\" : \"Invalid request parameters\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}